restore compare reloads checks
