+++ sdg/json.py
@@ -142,75 +142,3 @@ def write_json(inid, obj, ftype='data', gz=False, site_dir=''):
         return False
 
     return True
-
-
-
-# %% Compare reloads
-
-
-# This stuff needs to go in tests
-
-def isclose_df(df1, df2):
-    """A mix of np isclose and pandas equals that works across
-    python versions. So many api changes in pandas and numpy!"""
-    status = True
-    for col in df1:
-        if np.issubdtype(df1[col].dtype, np.number):
-            status = status & np.isclose(df1[col],
-                                         df2[col],
-                                         equal_nan=True).all()
-        else:
-            status = status & df1[col].equals(df2[col])
-    return status
-
-
-def compare_reload(inid, which='edges'):
-    """Load the original csv and compare to reloading the JSON you wrote out
-    which = 'edges' or 'data'
-    """
-    csv_path = indicator_path(inid, ftype=which, mode = 'w')
-
-    jsn = json.load(open(indicator_path(inid, 'json', mode = 'w')))
-
-    df_csv = pd.read_csv(csv_path, encoding='utf-8')
-    df_jsn = pd.DataFrame(jsn[which]).replace({None: np.nan})
-
-    # Account for empty data
-    if df_jsn.shape[0] == df_csv.shape[0] == 0:
-        return True
-
-    df_jsn = df_jsn[df_csv.columns.values]
-
-    status = isclose_df(df_csv, df_jsn)
-    if not status:
-        print("reload "+which+" error in "+inid)
-
-    return status
-
-# %% Read each csv and dump out to json
-
-def main():
-    """Read each csv and edge file and write out json. Then reload and 
-    check it's the same."""
-    status = True
-    # Create the place to put the files
-    os.makedirs("data/json", exist_ok=True)
-
-    ids = sdg.path.get_ids()
-    print("Building json for " + str(len(ids)) + " indicators...")
-
-    # For by record use orient='records'
-    # For column format use orient='list'
-    for inid in ids:
-        status = status & write_json(inid, orient='list', gz=False)
-        status = status & compare_reload(inid, 'data')
-        status = status & compare_reload(inid, 'edges')
-    return(status)
-
-
-if __name__ == '__main__':
-    status = main()
-    if(not status):
-        raise RuntimeError("Failed json dump")
-    else:
-        print("Success")
